+++ R/dependencies.R
@@ -1,6 +1,8 @@
 spark_dependencies <- function(spark_version, scala_version, ...) {
-  mleap_version <- if (spark_version >= "2.3.0") {
-    "0.12.0"
+  mleap_version <- if (spark_version >= "2.4.0") {
+    "0.14.0"
+  } else if (spark_version >= "2.3.0") {
+    "0.13.0"
   } else {
     "0.11.0"
   }
@@ -9,7 +11,8 @@ spark_dependencies <- function(spark_version, scala_version, ...) {
     jars = c(
       system.file(
         sprintf("java/mleap-%s-%s.jar", spark_version, scala_version),
-        package = "mleap"
+        package = "mleap",
+        mustWork = TRUE
       )
     ),
     packages = c(