+++ tests/testthat/test-serving.R
@@ -5,7 +5,7 @@ test_that("We can export and use pipeline model", {
   sc <- testthat_spark_connection()
   
   library(sparklyr)
-  mtcars_tbl <- sdf_copy_to(sc, mtcars, overwrite = TRUE)
+  mtcars_tbl <- copy_to(sc, mtcars, overwrite = TRUE)
   pipeline <- ml_pipeline(sc) %>%
     ft_binarizer("hp", "big_hp", threshold = 100) %>%
     ft_vector_assembler(c("big_hp", "wt", "qsec"), "features") %>%
@@ -15,7 +15,7 @@ test_that("We can export and use pipeline model", {
   # export model
   model_path <- file.path(tempdir(), "mtcars_model.zip")
   expect_message(ml_write_bundle(pipeline_model, 
-                                 ml_transform(pipeline_model, mtcars_tbl),
+                                 mtcars_tbl,
                                  model_path,
                                  overwrite = TRUE),
                  "Model successfully exported"
@@ -23,7 +23,7 @@ test_that("We can export and use pipeline model", {
   
   # error message when file exists
   expect_error(ml_write_bundle(pipeline_model, 
-                               ml_transform(pipeline_model, mtcars_tbl),
+                               mtcars_tbl,
                                model_path,
                                overwrite = FALSE),
                "*already exists\\.$"
@@ -58,8 +58,9 @@ test_that("We can export a list of transformers", {
   sc <- testthat_spark_connection()
   
   library(sparklyr)
-  iris_tbl <- sdf_copy_to(sc, iris, overwrite = TRUE)
-  string_indexer <- ft_string_indexer(sc, "Species", "label", dataset = iris_tbl)
+  iris_tbl <- copy_to(sc, iris, overwrite = TRUE)
+  string_indexer <- ft_string_indexer(sc, "Species", "label") %>% 
+    ml_fit(iris_tbl)
   pipeline <- ml_pipeline(string_indexer) %>%
     ft_vector_assembler(c("Petal_Width", "Petal_Length"), "features") %>%
     ml_logistic_regression() %>%
@@ -68,12 +69,10 @@ test_that("We can export a list of transformers", {
   pipeline_model <- ml_fit(pipeline, iris_tbl)
   stages <- pipeline_model %>%
     ml_stages(c("vector_assembler", "logistic", "index_to_string"))
-  transformed_tbl <- stages %>%
-    purrr::reduce(sdf_transform, .init = iris_tbl)
   model_path <- file.path(tempdir(), "iris_model.zip")
   
   expect_message(
-    ml_write_bundle(stages, transformed_tbl, model_path, overwrite = TRUE),
+    ml_write_bundle(stages, iris_tbl, model_path, overwrite = TRUE),
     "Model successfully exported"
   )
   
@@ -92,7 +91,7 @@ test_that("We can export a list of transformers", {
     5.2,          1.8
   )
   
-  newdata_tbl <- sdf_copy_to(sc, newdata, overwrite = TRUE)
+  newdata_tbl <- copy_to(sc, newdata, overwrite = TRUE)
   
   transformed_df <- mleap_transform(model, newdata)
   predictions_mleap <- transformed_df %>%
@@ -116,7 +115,7 @@ test_that("mleap_transform() handles heterogenous predictors", {
   skip_on_cran()
   sc <- testthat_spark_connection()
   
-  diamonds_tbl <- sdf_copy_to(sc, ggplot2::diamonds) %>%
+  diamonds_tbl <- copy_to(sc, ggplot2::diamonds) %>%
     dplyr::mutate(price = as.numeric(price))
   
   pipeline <- ml_pipeline(sc) %>%
@@ -136,7 +135,7 @@ test_that("mleap_transform() handles heterogenous predictors", {
   model_path <- file.path(tempdir(), "diamonds_model.zip")
   ml_write_bundle(
     pipeline_model,
-    ml_transform(pipeline_model, diamonds_tbl),
+    diamonds_tbl,
     model_path)
   
   mleap_model <- mleap_load_bundle(model_path)
@@ -147,7 +146,7 @@ test_that("mleap_transform() handles heterogenous predictors", {
     x = 4.5, y = 4.6, z = 2.7,
     stringsAsFactors = FALSE
   )
-  pred_data_tbl <- sdf_copy_to(sc, pred_data)
+  pred_data_tbl <- copy_to(sc, pred_data)
   
   prediction_mleap <- mleap_transform(mleap_model, pred_data) %>%
     dplyr::pull(prediction)
@@ -157,5 +156,27 @@ test_that("mleap_transform() handles heterogenous predictors", {
     dplyr::pull(prediction)
   
   expect_equal(prediction_mleap, prediction_spark)
+})
+
+test_that("Error when bundle doesn't have zip extension", {
+  skip_on_cran()
+  sc <- testthat_spark_connection()
   
+  library(sparklyr)
+  mtcars_tbl <- copy_to(sc, mtcars, overwrite = TRUE)
+  pipeline <- ml_pipeline(sc) %>%
+    ft_binarizer("hp", "big_hp", threshold = 100) %>%
+    ft_vector_assembler(c("big_hp", "wt", "qsec"), "features") %>%
+    ml_gbt_regressor(label_col = "mpg")
+  pipeline_model <- ml_fit(pipeline, mtcars_tbl)
+  
+  # export model
+  model_path <- file.path(tempdir(), "mtcars_model")
+
+  expect_error(ml_write_bundle(pipeline_model, 
+                               mtcars_tbl,
+                               model_path,
+                               overwrite = TRUE),
+               "The bundle path must have a `\\.zip` extension\\."
+  )
 })